Unit Testing for activity CRUD

+++ src/controllers/activity.controller.js
@@ -7,9 +7,23 @@ import * as logicSequenceCtrl from './logic_sequence.controller';
 
 //Get all activities from DB
 export const getActivities = async(req, res) => {
-    const activities = await Activity.find();
-    const count = await Activity.countDocuments();
-    res.status(200).json({message: "Activities list request has been completed satisfactorily", activities, count});
+    await Activity.find(async(err, activities) => {
+        if (err) {
+            console.log("ERROR when we trying to get all activities");
+            console.log(err);
+            return res.status(500).json({ message: "Unexpected error, try again later!"});
+        }
+        await Activity.countDocuments((error, count) => {
+            if (error) {
+                console.log("ERROR when we trying to get all activities");
+                console.log(error);
+                return res.status(500).json({ message: "Unexpected error, try again later!"})
+            }
+            return res.status(200).json({message: "Activities list request has been completed satisfactorily", activities, count});
+        });
+        
+    });
+    
 };
 
 //Create a new activity
@@ -17,69 +31,87 @@ export const createActivity = async(req, res) => {
     const { name, description, type } = req.body;
 
     //Verifying Fields
-    if (name == undefined) {
-        return res.status(400).json({message: "Field(s) required!"})
+    if (!name) {
+        return res.status(400).json({ message: "Field(s) required!" })
     }
+    if(type.localeCompare("logic_sequence") != 0 &&
+        type.localeCompare("maze") != 0 &&
+        type.localeCompare("questionnaire")) {
+            return res.status(400).json({ message: "Invalid type" })
+        }
 
     let temName = name.trim();
 
     //Verifying Fields
-    if(temName == undefined || temName.localeCompare("") == 0) {
-        return res.status(400).json({message: "Field(s) required!"})
+    if(temName.localeCompare("") == 0) {
+        return res.status(400).json({ message: "Field(s) required!" })
     }
 
     //Creating a new activity model
-    const newActivity = new Activity({name: temName, description, type})
+    const newActivity = new Activity({ name: temName, description, type })
 
     // Save the activity in the DB
-    const savedActivity = await newActivity.save();
+    await newActivity.save((err, savedActivity) =>{
+        if (err) return res.status(500).json({ message: "Unexpected error, try again later!"})
+        let child; 
+        if(type.localeCompare("logic_sequence") == 0) {
+            child = logicSequenceCtrl.createLogicSequence(savedActivity._id);
+        } else if(type.localeCompare("maze") == 0) {
+            console.log("maze");
+            //Do some stuff
+        } else if(type.localeCompare("questionnaire") == 0) {
+            console.log("questionnaire");
+            //Do some stuff
+        }
 
-    let child; 
-    if(type.localeCompare("logic_sequence") == 0) {
-        child = logicSequenceCtrl.createLogicSequence(savedActivity._id);
-    }
-
-    child.then((result) => {
-        res.status(201).json({message: "The activity has been created satisfactorily", activity: savedActivity, savedChild: result});
-    }).catch(err => {
-        console.log("ERROR found in createLogicSequence(logic_sequence.controller)")
-        console.err(err);
-        res.status(500).json({ message: "Unexpected error, try again later!"})
+        if(child) {
+            child.then(() => {
+                return res.status(201).json({message: "The activity has been created satisfactorily", activity_id: savedActivity._id});
+            }).catch(err => {
+                console.log("ERROR found in createLogicSequence(logic_sequence.controller)")
+                console.error(err);
+                return res.status(500).json({ message: "Unexpected error, try again later!"})
+            });
+        }
+        else {
+            return res.status(400).json({ message: "Type no accepted"});
+        }
     });
-
-    
 };
 
 //Update an activity
 export const updateActivityById = async(req, res) => {
 
-    const activity = await Activity.findById(req.params.id);
-    console.log("req.body.child")
-    console.log(req.body.child)
-    let child;
-    console.log("here")
-    if(activity.type.localeCompare("logic_sequence") == 0) {
-        child = logicSequenceCtrl.updateLogicSequenceByActivityId(activity._id, req.body.child);
-    }
-    
-    child.then(async(childResult) => {
-        console.log("CHILD11");
-        console.log(childResult);
-        await Activity.findByIdAndUpdate(req.params.id, req.body.activity, {
-            new: true
-        }).then(result => {
-            console.log("CHILD");
-            console.log(child);
-            res.status(201).json({ message: "The activity has been updated satisfactorily", updatedActivity: result, updatedChild: childResult });
+    await Activity.findById(req.params.id, (err, activity) => {
+        if(err) {
+            return res.status(500).json({ message: "Unexpected error, try again later!"});
+        }
+        let child;
+        if(activity.type.localeCompare("logic_sequence") == 0) {
+            child = logicSequenceCtrl.updateLogicSequenceByActivityId(activity._id, req.body.child);
+        } else if(type.localeCompare("maze") == 0) {
+            console.log("maze");
+            //Do some stuff
+        } else if(type.localeCompare("questionnaire") == 0) {
+            console.log("questionnaire");
+            //Do some stuff
+        }
+        if(!child) return res.status(500).json({ message: "Unexpected error, try again later!"});
+        child.then(async() => {
+            await Activity.findByIdAndUpdate(req.params.id, req.body.activity, {
+                new: true
+            }).then(() => {
+                return res.status(201).json({ message: "The activity has been updated satisfactorily"});
+            }).catch(err => {
+                console.log("ERROR found in updateActivityById(activity.controller)");
+                console.error(err);
+                return res.status(500).json({ message: "Unexpected error, try again later!"});
+            })
         }).catch(err => {
-            console.log("ERROR found in updateActivityById(activity.controller)")
-            console.err(err);
-            res.status(500).json({ message: "Unexpected error, try again later!"})
-        })
-    }).catch(err => {
-        console.log("ERROR found in updateLogicSequenceByActivityId(logic_sequence.controller)")
-        console.err(err);
-        res.status(500).json({ message: "Unexpected error, try again later!"})
+            console.log("ERROR found in updateLogicSequenceByActivityId(logic_sequence.controller)");
+            console.error(err);
+            return res.status(500).json({ message: "Unexpected error, try again later!"});
+        });
     });
 
     
@@ -88,15 +120,47 @@ export const updateActivityById = async(req, res) => {
 //Delete an activity
 export const deleteActivityById = async(req, res) => {
 
-    const activity = await Activity.findById(req.params.id);
-    let child;
-    if(activity.type.localeCompare("logic_sequence") == 0) {
-        child = await logicSequenceCtrl.deleteLogicSequenceById(activity._id);
-    }
+    await Activity.findById(req.params.id, (err, activity) => {
+        if(err) {
+            // console.log("ERROR when try to findById in deleteActivityById (activity.controller)");
+            // console.error(err);
+            return res.status(500).json({ message: "Unexpected error, try again later!"});
+        }
 
-    await Activity.deleteOne({_id: activity._id}, function(err) {
-        if (err) return handleError(err);
+        if(!activity) {
+            return res.status(400).json({ message: "Activity not found"});
+        }
+
+        let child;
+        if(activity.type.localeCompare("logic_sequence") == 0) {
+            child = logicSequenceCtrl.deleteLogicSequenceById(activity._id);
+        } else if(type.localeCompare("maze") == 0) {
+            console.log("maze");
+            //Do some stuff
+        } else if(type.localeCompare("questionnaire") == 0) {
+            console.log("questionnaire");
+            //Do some stuff
+        }
+
+        if(child) {
+            child.then(async() => {
+                await Activity.deleteOne({_id: activity._id}, function(err) {
+                    if(err) {
+                        console.error("ERROR when try to deleteOne in deleteActivityById (activity.controller)");
+                        return res.status(500).json({ message: "Unexpected error, try again later!"})
+                    }
+
+                    return res.status(200).json({message: "The activity has been deleted satisfactorily"});
+                });
+            }).catch((err) => {
+                console.log("ERROR found in deleteActivityById(activity.controller)")
+                console.error(err);
+                return res.status(500).json({ message: "Unexpected error, try again later!"})
+            });
+        } else {
+            return res.status(500).json({ message: "Unexpected error, try again later!"})
+        }
     });
-
-    res.status(200).json({message: "The activity has been deleted satisfactorily", deletedActivity: activity, deletedChild: child});
+  
+    
 };
